requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
deprecation==2.1.0
exceptiongroup==1.3.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.7.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
websockets==15.0.1
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
rpi-lgpio==0.6
